python/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
PyPDF2>=2.0.0
python-docx>=0.8.11
openpyxl>=3.0.7
pdfminer.six>=20200726
